Updated Task3 with inheritance from super class
